Move fixtures to local conftest
